+++ openfisca_france/model/mesures.py
@@ -200,9 +200,9 @@ class plus_values_revenus_nets_du_capital(Variable):
         mais pour lesquelles on n'a pas de prélèvements sociaux
         Cette variable est l'assiette de plus-values pour lesquelles au moins un prélèvement est calculé. On l'utilise dans le
         calcul du revenu disponible, afin de n'oublier aucun revenu. Elle vaut la somme de assiette_csg_plus_values et rfr_plus_values_hors_rni,
-        où l'on enlève les cases communes entre ces deux variables, et où l'on ajoute les variables présentes dans rev_cat_pv, mais pas présente
+        où l'on enlève les cases communes entre ces deux variables, et où l'on ajoute les variables présentes dans 'revenu_categoriel_plus_values', mais pas présente
         dans assiette_csg_plus_values
-        Attention : pour les variables de rev_cat_pv ajoutées, elles peuvent représenter des montants nets, alors qu'il faudrait le brut. Améliorer ce point
+        Attention : pour les variables de 'revenu_categoriel_plus_values' ajoutées, elles peuvent représenter des montants nets, alors qu'il faudrait le brut. Améliorer ce point
         '''
 
         v1_assiette_csg_plus_values = foyer_fiscal('assiette_csg_plus_values', period)
@@ -214,7 +214,7 @@ class plus_values_revenus_nets_du_capital(Variable):
 
         intersection_v1_v2 = f3vg + f3we + f3vz
 
-        return v1_assiette_csg_plus_values + v2_rfr_plus_values_hors_rni - intersection_v1_v2  # Pas d'item supplémentaire dans rev_cat_pv avant 2013
+        return v1_assiette_csg_plus_values + v2_rfr_plus_values_hors_rni - intersection_v1_v2
 
     def formula_2013_01_01(foyer_fiscal, period):
         '''
@@ -261,12 +261,15 @@ class plus_values_revenus_nets_du_capital(Variable):
         v1_assiette_csg_plus_values = foyer_fiscal('assiette_csg_plus_values', period)
         v2_rfr_plus_values_hors_rni = foyer_fiscal('rfr_plus_values_hors_rni', period)
 
-        f3vg = foyer_fiscal('f3vg', period)
+        f3sg = foyer_fiscal('f3sg', period)
+        f3sl = foyer_fiscal('f3sl', period)
         f3ua = foyer_fiscal('f3ua', period)
+        f3vg = foyer_fiscal('f3vg', period)
+        f3vz = foyer_fiscal('f3vz', period)
 
-        intersection_v1_v2 = f3vg + f3ua
+        intersection_v1_v2 = f3vg + f3ua + f3vz + f3sg + f3sl
 
-        return v1_assiette_csg_plus_values + v2_rfr_plus_values_hors_rni - intersection_v1_v2  # Pas d'item supplémentaire dans rev_cat_pv avant 2013
+        return v1_assiette_csg_plus_values + v2_rfr_plus_values_hors_rni - intersection_v1_v2
 
 
 class revenus_nets_du_capital(Variable):
@@ -298,7 +301,7 @@ class revenus_nets_du_capital(Variable):
         assiette_csg_revenus_capital = foyer_fiscal('assiette_csg_revenus_capital', period)
         assiette_csg_plus_values = foyer_fiscal('assiette_csg_plus_values', period)
         plus_values_revenus_nets_du_capital = foyer_fiscal('plus_values_revenus_nets_du_capital', period)
-        rev_cat_rfon = foyer_fiscal('rev_cat_rfon', period)
+        rev_cat_rfon = foyer_fiscal('revenu_categoriel_foncier', period)
         rente_viagere_titre_onereux_net = foyer_fiscal('rente_viagere_titre_onereux_net', period)
         fon = foyer_fiscal('fon', period)